Lazy-load startup shell paths
