+++ src/features/catalog/agents/debug-shell.ts
@@ -1,0 +1,1 @@
+export { AgentDebugPanel } from "./components/AgentDebugPanel";
