+++ src/features/catalog/chats/sidebar.ts
@@ -1,0 +1,12 @@
+export { useBulkExportChats } from "./hooks/use-bulk-export-chats";
+export { useChatSummaries } from "./hooks/use-chat-summaries";
+export { useCreateChat, useDeleteChat, useDeleteChatGroup, useUpdateChatMetadata } from "./hooks/use-chat-lifecycle";
+export {
+  useChatFolders,
+  useCreateFolder,
+  useDeleteFolder,
+  useMoveChat,
+  useReorderFolders,
+  useUpdateFolder,
+} from "./hooks/use-chat-folders";
+export type { BulkChatExportFormat } from "./lib/chat-transcript-export";
